# lib/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  title: Api
  version: "1.0.0"
servers:
  - url: /api
paths:
...
  /operators:
    get:
      operationId: getOperators
      summary: List all operators
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Operator"
    post:
      operationId: createOperator
      summary: Create operator (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperatorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operator"
  /operators/{id}:
    get:
      operationId: getOperator
      summary: Get operator by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operator"
    put:
      operationId: updateOperator
      summary: Update operator (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperatorInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operator"
    delete:
      operationId: deleteOperator
      summary: Delete operator (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Stations ─────────────────────────────────────────────────
...
  /routes:
    get:
      operationId: getRoutes
      summary: List routes
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Route"
    post:
      operationId: createRoute
      summary: Create a smart route
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
  /routes/{id}:
    get:
      operationId: getRoute
      summary: Get route by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
    delete:
      operationId: deleteRoute
      summary: Delete route
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Charging Sessions ────────────────────────────────────────
...
  /payment-methods:
    get:
      operationId: getPaymentMethods
      summary: List payment methods for user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentMethod"
    post:
      operationId: addPaymentMethod
      summary: Add payment method
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentMethodInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentMethod"
  /payment-methods/{id}:
    delete:
      operationId: deletePaymentMethod
      summary: Remove payment method
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Admin Auth ───────────────────────────────────────────────
  /admin/login:
    post:
      operationId: adminLogin
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminToken"
        "401":
          description: Unauthorized

  # ── Admin Dashboard ──────────────────────────────────────────
...
components:
  schemas:
...
    Operator:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        logo_url:
          type: ["string", "null"]
        station_count:
          type: integer
    OperatorInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        logo_url:
          type: ["string", "null"]

    # ── Connector ─────────────────────────────────────────────
...
    Vehicle:
      type: object
      required: [id, name, battery_kwh, range_km, connector_type]
      properties:
        id:
          type: integer
        name:
          type: string
        battery_kwh:
          type: number
        range_km:
          type: number
        connector_type:
          type: string
          enum: [CCS2, CHAdeMO, Type2, GB-T]
        current_battery_pct:
          type: ["number", "null"]
...
    RouteStop:
      type: object
      required: [station_id, station_name, arrival_battery_pct, departure_battery_pct, charge_time_min, distance_from_prev_km]
      properties:
        station_id:
          type: integer
        station_name:
          type: string
        address:
          type: string
        lat:
          type: number
        lng:
          type: number
        arrival_battery_pct:
          type: number
        departure_battery_pct:
          type: number
        charge_time_min:
          type: integer
        distance_from_prev_km:
          type: number
        eta:
          type: string

    # ── Route ─────────────────────────────────────────────────
    Route:
      type: object
      required: [id, origin, destination, total_distance_km, total_time_min, status, created_at]
      properties:
        id:
          type: integer
        vehicle_id:
          type: ["integer", "null"]
        vehicle:
          $ref: "#/components/schemas/Vehicle"
        origin:
          type: string
        destination:
          type: string
        origin_lat:
          type: ["number", "null"]
        origin_lng:
          type: ["number", "null"]
        dest_lat:
          type: ["number", "null"]
        dest_lng:
          type: ["number", "null"]
        initial_battery_pct:
          type: number
        stops:
          type: array
          items:
            $ref: "#/components/schemas/RouteStop"
        total_distance_km:
          type: number
        total_time_min:
          type: integer
        status:
          type: string
          enum: [active, completed, cancelled]
        created_at:
          type: string
          format: date-time
    RouteInput:
      type: object
      required: [origin, destination, initial_battery_pct]
      properties:
        origin:
          type: string
        destination:
          type: string
        origin_lat:
          type: ["number", "null"]
        origin_lng:
          type: ["number", "null"]
        dest_lat:
          type: ["number", "null"]
        dest_lng:
          type: ["number", "null"]
        vehicle_id:
          type: ["integer", "null"]
        initial_battery_pct:
          type: number

    # ── Session ───────────────────────────────────────────────
...
    PaymentMethod:
      type: object
      required: [id, type, last_four, is_default]
      properties:
        id:
          type: integer
        user_id:
          type: string
        type:
          type: string
        last_four:
          type: string
        is_default:
          type: boolean
    PaymentMethodInput:
      type: object
      required: [type, last_four]
      properties:
        user_id:
          type: ["string", "null"]
        type:
          type: string
        last_four:
          type: string
        is_default:
          type: boolean

    # ── Admin ─────────────────────────────────────────────────
    AdminLoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AdminToken:
      type: object
      required: [token, email]
      properties:
        token:
          type: string
        email:
          type: string
...