# lib/api-zod/src/generated/types/operator.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface Operator {
  id: number;
  name: string;
  /** @nullable */
  logo_url?: string | null;
  station_count?: number;
}