# lib/api-zod/src/generated/types/operatorInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface OperatorInput {
  name: string;
  /** @nullable */
  logo_url?: string | null;
}