# lib/api-zod/src/generated/types/paymentMethod.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface PaymentMethod {
  id: number;
  user_id?: string;
  type: string;
  last_four: string;
  is_default: boolean;
}