# lib/api-zod/src/generated/types/route.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
import type { RouteStatus } from './routeStatus';
import type { RouteStop } from './routeStop';
import type { Vehicle } from './vehicle';

export interface Route {
  id: number;
  /** @nullable */
  vehicle_id?: number | null;
  vehicle?: Vehicle;
  origin: string;
  destination: string;
  /** @nullable */
  origin_lat?: number | null;
  /** @nullable */
  origin_lng?: number | null;
  /** @nullable */
  dest_lat?: number | null;
  /** @nullable */
  dest_lng?: number | null;
  initial_battery_pct?: number;
  stops?: RouteStop[];
  total_distance_km: number;
  total_time_min: number;
  status: RouteStatus;
  created_at: Date;
}