# lib/api-zod/src/generated/types/routeInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface RouteInput {
  origin: string;
  destination: string;
  /** @nullable */
  origin_lat?: number | null;
  /** @nullable */
  origin_lng?: number | null;
  /** @nullable */
  dest_lat?: number | null;
  /** @nullable */
  dest_lng?: number | null;
  /** @nullable */
  vehicle_id?: number | null;
  initial_battery_pct: number;
}