# lib/api-zod/src/generated/types/routeStop.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface RouteStop {
  station_id: number;
  station_name: string;
  address?: string;
  lat?: number;
  lng?: number;
  arrival_battery_pct: number;
  departure_battery_pct: number;
  charge_time_min: number;
  distance_from_prev_km: number;
  eta?: string;
}